requirements.txt
appdirs==1.4.4
asgiref==3.4.0
astroid==2.2.5
autopep8==1.5.1
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.3.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.1.0
importlib-metadata==3.7.3
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pep8==1.7.1
pipenv==2020.11.15
psycopg2==2.7.7
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dotenv==0.18.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
virtualenv==20.4.3
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.11.2
zipp==3.4.1
